Bugged ver for Rex to examine

+++ develop/GEN1_HYBRID_MAINBOARD/src/user/auto_mode.h
@@ -33,7 +33,7 @@ typedef	struct {
 	int x;
 	int y;
 	int deg;
-	double curve;
+	int curve;
 } TARGET;
 
 #include "auto_paths.h"
